ima/pyproject.toml
[project]
name = "ima"
version = "0.1.0"
description = "IMA permission-based auth service (FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio[asyncio,trio]>=4.10.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "redis[async]>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pre-commit-hooks>=6.0.0",
    "pylint>=3.3.8",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.15.0",
    "ruff>=0.12.12",
]


[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
fix = true
exclude = ["alembic"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
